Start implementing the Family Trainer mat
Reading the controller is currently implemented and works on hardware.
Only Side B is currently displayed for button inputs.

Next on the agenda is figuring out the wait timing between controller
reads.  Currently, about 1800 cycles are burned between selecting a row
by writing to $4016 and reading the data from $4017.  The Family Basic
keyboard reportedly only needs ~50 cycles, so it might be similar for
the Family Trainer.