Implement variable delay reading Family Trainer
Implemented a variable delay time when reading button presses from the
Family Trainer.  This delay comes between the write to $4016 and the
read from $4017.  If this delay is too short, additional button presses
will register (typically other buttons in the same column).

This delay value is displayed on screen in hexadecimal.  The value is
not in cycles, but in loop itterations.  Each loop is 19 cycles long
(minus 1 for the final itteration), plus the prep and teardown of the
routine totaling another 17 cycles.

The Left and B buttons lower the delay value while Right and A increase
the delay.  There is no wrap around checking on the value.  A value of
$FF will make the controller read routine take roughly half of a frame
to execute.

In my testing, a total of about 758 cycles (39 loops) between the write
to $4016 and the read from $4017 seemed to be the lowest reliable delay
without ghost inputs.  In real world applications, a longer delay should
probably be employed as there is no harm in waiting longer.