Finish adding titler stuff
Read the touchpad and display the raw values as well as a representation
of the location with a sprite.

This code has been successfully verified on hardware, but will softlock
in emulators.  Reading the touchpad requires the use of a hardware
generated IRQ that no emulator currently generates.

Verifying the touchpad on hardware required disabling the IRQ pin on my
Everdrive via an adapter.  Loading this rom on to an Everdrive and using
that directly will softlock the screen.  Additionally, the switch on the
side of the console labeled 文字入力 (character entry) must be set to 入
(on/enter) in order to read either the touchpad or front pannel buttons.